Handle missing image url and remove extra debug print statements
The new method of picking info from RPC calls is more sensitive to posts
missing json metadata. In this case, there's no image found. As a stop gap
I've caught an exception on it not being found then set the image to blank
for manual fixing. The real bugfix will be to try to infer the image
thumbnail in another manner - probably just grab it the 'old way' via
webscraping.

As part of tracking this down, a bunch of debug prints were commented out.